rpi/requirements_rpi.txt
onnxruntime
torchvision
Pillow